fix(merge.ts): clone route argument to avoid mutation
This uses `Object.assign` to clone the `route` argument before assigning
it to the `options` variable.

Before this change a headers field was being set on the `options`
variable, which was the same reference. Another effect of this change is
that the headers field on the `route` argument would be mutated to be
lowercase. Now the lowercase operation happens on the cloned value
instead of mutating the argument passed in to the function.