fix: fix support for Node.js v18's native `fetch` when making `DELETE` requests with no body
Currently, `DELETE` requests without a body don't work on the
latest Node.js version, v18, because we try to set
`content-length: 0` which is not allowed by the native `fetch`
implementation which is included in Node. It throws an error
with the `UND_ERR_REQ_CONTENT_LENGTH_MISMATCH` code.

This stops manually setting that `content-length: 0` header
for `DELETE` requests.

I've tested this against the GitHub API with Node.js v18
and v17, and both work as expected, with the GitHub API also
accepting the request.

As part of this PR, I've also added some new tests covering
what happens with different HTTP methods.

Fixes https://github.com/octokit/request.js/issues/506.